Refactor requirements validation: split logic between network bootstrap and kube-proxy management.
