+++ pkg/crds/operator/operator.tigera.io_installations.yaml
@@ -1277,7 +1277,7 @@ spec:
                     bpfNetworkBootstrap:
                       description: |-
                         BPFNetworkBootstrap defines how the initial setup for configuring the BPF dataplane is performed.
-                        When enabled, the operator automatically includes the API server and EndpointSlice addresses in the NAT maps.
+                        When enabled, the Kubernetes service and associated endpoints will be used to bootstrap access to the kubernetes API server.
                         Default: Disabled
                       enum:
                         - Disabled
@@ -10066,7 +10066,7 @@ spec:
                         bpfNetworkBootstrap:
                           description: |-
                             BPFNetworkBootstrap defines how the initial setup for configuring the BPF dataplane is performed.
-                            When enabled, the operator automatically includes the API server and EndpointSlice addresses in the NAT maps.
+                            When enabled, the Kubernetes service and associated endpoints will be used to bootstrap access to the kubernetes API server.
                             Default: Disabled
                           enum:
                             - Disabled
